api/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
torch==2.0.1
torchvision==0.15.2
torchaudio==2.0.2
pillow>=9.5.0,<10
numpy<2
opencv-python>=4.7.0.72,<4.8
prometheus-client==0.17.0
ultralytics>=8.0.0
seaborn==0.12.2
matplotlib==3.7.1
python-multipart
tqdm==4.66.1
mlflow
prometheus-fastapi-instrumentator
